+++ RELEASES.md
@@ -5,6 +5,7 @@
 ### Added
 
 * ibmpowernv: new implementation for IBM PowerNV systems (targeted for OLCF Summit)
+* ipg: new implementation using the Intel Power Gadget 3.7 API (PowerGadgetLib.h)
 
 ### Fixed